Bug 1262389 - Split kinto-client.js
MozReview-Commit-ID: KxUxoXEnn05
