+++ services/common/moz.build
@@ -16,6 +16,7 @@ EXTRA_COMPONENTS += [
 EXTRA_JS_MODULES['services-common'] += [
     'async.js',
     'kinto-updater.js',
+    'kinto-http-client.js',
     'KintoCertificateBlocklist.js',
     'logmanager.js',
     'moz-kinto-client.js',
